requirements.txt
django==2.2
djangorestframework==3.11.0
markdown==3.2.2
factory-boy==2.12.0
django-extensions==2.2.9